Working version of prealignment
I have been having a lot of problems with running the prealignment. I
have added several things to get things running smooth. First I added
`-k` to the WRAPPER_SLURM script so that even if a job fail it continues
to run all of the other jobs. I thought this was the default behavior
but I was wrong.

I also ended up installing my own sratoolkit because I was not sure if
my settings were being overridden by biowulf's configuration. This may
not be necessary.

I also reduced the `-j` parameter to 100. There are a lot of jobs that
fail because of SRA's not being downloaded correctly. I am not sure if
this is an SRA problem or a biowulf problem. I know there are some
restriction when querying the ENTREZ using E-utilities, these limits may
also apply to sratoolkit. Biowulf uses a proxy connection to the SRA, so
perhaps hitting this proxy with too many requests may be problematic.

There are still a number of jobs failing, I will have to go back and
trying re-running these jobs once the big job is done.