Adds additional logic to set as SE if R1 != R2.
I have found that some of the uploaded files have been trimmed and
filtered. If I don't have matching mates, this causes havoc in
downstream programs. I could approach this problem three ways.

1. Remove samples that do not have even number of reads.
2. Use R1 (in case stranded) and ignore R2.
3. Try to filter out reads and keep those that have both mates.

I have decided to go with (2). (1) does not seem to be a good options,
and (3) has a lot of potential problems related to possibly removing
most of the reads. I think (2) is the best option for simplicity and
usefulness.