Change workflow to merge at the end.
For some silly reason I was trying to merge bams and then do the
coverages. This was taking forever to run, so I decided to keep with the
map-reduce mentality and run everything separately and combine at the
end. This requires some additional steps, but seemed to make more sense
in the end and should be faster to run.