Removes a bunch of depricated files.
Space on biowulf was really tight and I realized that there were a lot
of files that had not been garbage collected because of iterations of
the pipeline.

This notebook goes through the completed runs and checks for files
needing garbage collected. Runs not in the complete list should be
garbage collected as part of the pipeline.