Updates parser calls.
I had updated the parsers to only take a filename. I was trying to
simplify the code a little. This broke the calls in the workflow, so I
just change them to use the updated function calls and pass a file
directly instead of passing a pattern and the fill options.