S2 cell rna-seq.
For the S2rnai project I need a list of S2 cell RNA-seq samples from the
SRA. Here I use the database to pull out putative samples. I then dumped
to excel and hand edited the list and re-imported it. Finally I munge
the counts into a single table to send to Yijie.